move createTableDataSourceConfiguration to configuration
